pyproject.toml
[project]
name = "websiteupdater"
version = "0.1.0"
description = "Automated dependency-update-and-verify pipeline for Laravel (and PHP/Node) projects deployed on Laravel Cloud."
requires-python = ">=3.11"
readme = "README.md"
authors = [{ name = "Your Name", email = "you@example.com" }]

dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "playwright>=1.44",
    "pixelmatch>=0.3.0",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
wu = "websiteupdater.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["websiteupdater*"]
